backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.3.0
pdfplumber==0.11.0
openfoodfacts==1.1.0
httpx==0.27.0
python-dotenv==1.0.1
